feat(auth, docs): update the docs and complete flow for reset password
create a 2 routes for reset password, the first is reset-password here the user put's his email and
supabase sent a link to redirecto user to second route update-password where the user change his
password.
