--- src/app/actions.ts
@@ -1,19 +1,0 @@
-'use server'
-
-import { revalidatePath } from 'next/cache'
-import { redirect } from 'next/navigation'
-
-import { createClient } from '@/utils/supabase/server'
-
-export async function logout() {
-  const supabase = createClient()
-
-  const { error } = await supabase.auth.signOut()
-
-  if (error) {
-    redirect('/?message=Error on signout')
-  }
-
-  revalidatePath('/', 'layout')
-  redirect('/auth/login')
-}
