+++ src/app/auth/update-password/components/UpdatePassword.tsx
@@ -1,0 +1,6 @@
+'use client'
+export default function updatePassword() {
+  return (
+    <div>updatePassword</div>
+  )
+}
