+++ src/components/Navbar.tsx
@@ -1,44 +1,40 @@
-import React from 'react'
-import { Button } from './ui/button'
 import { User } from '@supabase/supabase-js'
 import Link from 'next/link'
+import Alert from './Alert'
+import Logout from './Logout'
 
-export default function Navbar({
-  user,
-  logout,
-}: {
-  user: User
-  logout: () => {}
-}) {
+export default function Navbar({ user }: { user: User }) {
   const routes = [
     {
       label: 'Mis Libros',
-      path: '/account/my-books',
+      path: '/my-books',
     },
     {
       label: 'Añadir Libro',
-      path: '/account/add-book',
+      path: '/add-book',
     },
     {
       label: 'Notificaciónes',
-      path: '/account/notifications',
+      path: '/notifications',
     },
   ]
   return (
-    <nav className="fixed inset-x-0 top-0 flex h-[80px] w-full items-center justify-between border-b-2 border-gray-300 px-8">
-      <h1 className="text-xl font-extrabold uppercase">Librito facil 📚</h1>
-      <ul className="flex items-center gap-4 font-bold uppercase text-neutral-900">
-        {routes.map((route) => (
-          <li key={route.path}>
-            <Link href={route.path}>{route.label}</Link>
-          </li>
-        ))}
+    <>
+      <nav className="fixed inset-x-0 top-0 flex h-[80px] w-full items-center justify-between border-b-2 border-gray-300 px-8">
+        <h1 className="text-xl font-extrabold uppercase">Libro Baratito! 📚</h1>
+        <ul className="flex items-center gap-4 font-bold uppercase text-neutral-900">
+          {routes.map((route) => (
+            <li key={route.path}>
+              <Link href={route.path}>{route.label}</Link>
+            </li>
+          ))}
 
-        <li>{user.email}</li>
-        <form>
-          <Button formAction={logout}>Salir</Button>
-        </form>
-      </ul>
-    </nav>
+          <li>{user.email}</li>
+          <Logout />
+        </ul>
+      </nav>
+
+      <Alert />
+    </>
   )
 }
